Refactor : 수정 security 내부 패키지
security 패키지 내부 요소 리팩토링
