+++ application/src/main/java/core/application/filter/CustomLogoutFilter.java
@@ -2,7 +2,7 @@ package core.application.filter;
 
 import core.application.api.response.ApiResponse;
 import core.application.api.response.code.Message;
-import core.application.security.service.TokenService;
+import core.application.security.token.TokenService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.ServletRequest;
