+++ application/src/main/java/core/application/reviews/controllers/ReviewCommentController.java
@@ -7,7 +7,7 @@ import core.application.reviews.models.dto.request.*;
 import core.application.reviews.models.dto.response.comments.*;
 import core.application.reviews.models.entities.*;
 import core.application.reviews.services.*;
-import core.application.security.service.*;
+import core.application.security.auth.CustomUserDetails;
 import io.swagger.v3.oas.annotations.*;
 import io.swagger.v3.oas.annotations.tags.*;
 import jakarta.servlet.http.*;
