+++ application/src/main/java/core/application/users/controller/DibController.java
@@ -1,7 +1,7 @@
 package core.application.users.controller;
 
 import core.application.api.response.ApiResponse;
-import core.application.security.service.CustomUserDetails;
+import core.application.security.auth.CustomUserDetails;
 import core.application.users.models.dto.DibRespDTO;
 import core.application.users.service.DibService;
 import io.swagger.v3.oas.annotations.Operation;
